+++ docs/CHANGELOG.md
@@ -9,6 +9,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [dev]
 
+### Changed
+
+- `L2Suggested` mode is now called `SuggestedPrice`
+
+### Removed
+
+- Removed `Optimism2` as a supported gas estimator mode
+
 ...
 
 ## 2.7.0 - UNRELEASED